maintainer/requirements.txt
pyyaml>=5.4
cassandra-driver==3.29.0
wheel
pytest
